+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- [#10](https://github.com/RchrdHndrcks/gochess/pull/10) Fix promotion moves in capture.
+
 ## [1.0.3] - 2024-07-15
 
 ## Added